Small improvements to the Laraval docs
Call out the platform dependent value of the CA root chain and refer to
our documentation for values on different platforms.

Also update client certificates to call it the PlanetScale proxy which
is the better description for that usage. The fact that they use client
certificates is an implementation detail and not the name of how to
connect with it.