Fix broken image source for query insights list
When you scroll down to this part of the [Concepts > Billing](https://planetscale.com/docs/concepts/billing) page, there's an up and down jitter in Firefox due to repeated failing attempts to load the the queries list screenshot. I noticed that the image URL is probably supposed to be `docs/concepts/query-insights/queries.png`.