Improve deploy deploy requests GitHub Action
If a PR is closed but not merge, you probably don't want to deploy a deploy request (usually). This adds a check to the sample config to make sure that PRs are merged.

Also updates the example and parameter based on the latest [v4 readme](https://github.com/planetscale/deploy-deploy-request-action#readme).